function_app/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-identity>=1.17.0
azure-keyvault-secrets>=4.8.0
azure-eventhub>=5.12.0
faker
python-dateutil